% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomalies.R
\name{detect_anomalies}
\alias{detect_anomalies}
\title{Detect anomalies generic}
\usage{
detect_anomalies(obj, ...)
}
\arguments{
\item{obj}{Object to detect anomalies on}

\item{...}{Additional arguments}
}
\description{
Generic function for detecting anomalies in different objects.
}
