% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_anomalies}
\alias{plot_anomalies}
\title{Plot anomalies for a deposit dataset}
\usage{
plot_anomalies(obj)
}
\arguments{
\item{obj}{A \code{deposit_dataset} object with anomaly results}
}
\value{
A ggplot2 object (or list of ggplot2 objects)
}
\description{
Plot anomalies for a deposit dataset
}
