% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.deposit_dataset}
\alias{plot.deposit_dataset}
\title{Plot monthly anomalies for a deposit_dataset object}
\usage{
\method{plot}{deposit_dataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{deposit_dataset} with anomalies detected.}
}
\description{
Plot monthly anomalies for a deposit_dataset object
}
