% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.deposit_dataset}
\alias{summary.deposit_dataset}
\title{Summarize a deposit dataset}
\usage{
\method{summary}{deposit_dataset}(obj, excel_path = NULL)
}
\arguments{
\item{object}{An object of class \code{deposit_dataset}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints and exports a summary table of anomalies per customer.
}
\description{
Summarize a deposit dataset
}
