% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\docType{data}
\name{transactions}
\alias{transactions}
\title{Transactions dataset}
\format{
A data frame with columns:
\describe{
\item{customer_id}{Unique ID for each customer}
\item{date}{Transaction date (Date class)}
\item{amount}{Transaction amount (numeric)}
}
}
\usage{
transactions
}
\description{
Example dataset of deposit transactions.
}
\examples{
data(transactions)
}
\keyword{datasets}
